+++ Sources/PayKit/CustomerRequest+Extensions.swift
@@ -1,0 +1,26 @@
+//
+//  File.swift
+//
+//  Licensed under the Apache License, Version 2.0 (the "License");
+//  you may not use this file except in compliance with the License.
+//  You may obtain a copy of the License at
+//
+//  http://www.apache.org/licenses/LICENSE-2.0
+//
+//  Unless required by applicable law or agreed to in writing, software
+//  distributed under the License is distributed on an "AS IS" BASIS,
+//  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+//  See the License for the specific language governing permissions and
+//  limitations under the License.
+//
+
+import Foundation
+
+extension CustomerRequest.AuthFlowTriggers {
+    // Jitter to account for the requests that are about to expire
+    private static let expiryJitter: TimeInterval = 10
+
+    func isExpired(on date: Date = Date()) -> Bool {
+        refreshesAt <= date.addingTimeInterval(CustomerRequest.AuthFlowTriggers.expiryJitter)
+    }
+}